third_party/WebKit/LayoutTests/accessibility/br-element-has-correct-title-expected.txt

A br element
should return newline as its accessible title.

This tests that a br element will return the newline character as its accessible title.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS brElement.title is 'AXTitle: \n'
PASS successfullyParsed is true

TEST COMPLETE
